Remove QueryPriority
Motivation
----------
According to the Analytics team, setting the
"Analytics-Priority" header doesn't do much.
Notably, it does not affect the resources
allocated to a query.

The server team is working on a better long-term
solution for prioritizing short-lived and/or
high-priority queries.

Rather than confuse users with an option that
does not do what the name suggests, remove it for now,
and revisit when the server-side solution is complete.